refactor: Split out BUFFER-handling functions
