+++ init.c
@@ -21,6 +21,7 @@
 #endif
 
 #include "mutt.h"
+#include "filter.h"
 #include "mapping.h"
 #include "mutt_curses.h"
 #include "mutt_menu.h"
@@ -45,6 +46,7 @@
 #include "mx.h"
 #include "init.h"
 #include "mailbox.h"
+#include "myvar.h"
 
 #include <ctype.h>
 #include <stdlib.h>
@@ -73,10 +75,7 @@ typedef struct myvar
 
 static myvar_t* MyVars;
 
-static int var_to_string (int idx, char* val, size_t len);
-
 static void myvar_set (const char* var, const char* val);
-static const char* myvar_get (const char* var);
 static void myvar_del (const char* var);
 
 #if USE_NOTMUCH
@@ -133,7 +132,7 @@ int query_quadoption (int opt, const char *prompt)
 
 /* given the variable ``s'', return the index into the rc_vars array which
    matches, or -1 if the variable is not found.  */
-static int mutt_option_index (char *s)
+int mutt_option_index (char *s)
 {
   int i;
 
@@ -141,203 +140,6 @@ static int mutt_option_index (char *s)
     if (mutt_strcmp (s, MuttVars[i].option) == 0)
       return (MuttVars[i].type == DT_SYN ?  mutt_option_index ((char *) MuttVars[i].data) : i);
   return (-1);
-}
-
-int mutt_extract_token (BUFFER *dest, BUFFER *tok, int flags)
-{
-  char		ch;
-  char		qc = 0; /* quote char */
-  char		*pc;
-
-  /* reset the destination pointer to the beginning of the buffer */
-  dest->dptr = dest->data;
-
-  SKIPWS (tok->dptr);
-  while ((ch = *tok->dptr))
-  {
-    if (!qc)
-    {
-      if ((ISSPACE (ch) && !(flags & MUTT_TOKEN_SPACE)) ||
-	  (ch == '#' && !(flags & MUTT_TOKEN_COMMENT)) ||
-	  (ch == '=' && (flags & MUTT_TOKEN_EQUAL)) ||
-	  (ch == ';' && !(flags & MUTT_TOKEN_SEMICOLON)) ||
-	  ((flags & MUTT_TOKEN_PATTERN) && strchr ("~%=!|", ch)))
-	break;
-    }
-
-    tok->dptr++;
-
-    if (ch == qc)
-      qc = 0; /* end of quote */
-    else if (!qc && (ch == '\'' || ch == '"') && !(flags & MUTT_TOKEN_QUOTE))
-      qc = ch;
-    else if (ch == '\\' && qc != '\'')
-    {
-	if (!*tok->dptr)
-	    return -1; /* premature end of token */
-      switch (ch = *tok->dptr++)
-      {
-	case 'c':
-	case 'C':
-	    if (!*tok->dptr)
-		return -1; /* premature end of token */
-	  mutt_buffer_addch (dest, (toupper ((unsigned char) *tok->dptr)
-                                    - '@') & 0x7f);
-	  tok->dptr++;
-	  break;
-	case 'r':
-	  mutt_buffer_addch (dest, '\r');
-	  break;
-	case 'n':
-	  mutt_buffer_addch (dest, '\n');
-	  break;
-	case 't':
-	  mutt_buffer_addch (dest, '\t');
-	  break;
-	case 'f':
-	  mutt_buffer_addch (dest, '\f');
-	  break;
-	case 'e':
-	  mutt_buffer_addch (dest, '\033');
-	  break;
-	default:
-	  if (isdigit ((unsigned char) ch) &&
-	      isdigit ((unsigned char) *tok->dptr) &&
-	      isdigit ((unsigned char) *(tok->dptr + 1)))
-	  {
-
-	    mutt_buffer_addch (dest, (ch << 6) + (*tok->dptr << 3) + *(tok->dptr + 1) - 3504);
-	    tok->dptr += 2;
-	  }
-	  else
-	    mutt_buffer_addch (dest, ch);
-      }
-    }
-    else if (ch == '^' && (flags & MUTT_TOKEN_CONDENSE))
-    {
-	if (!*tok->dptr)
-	    return -1; /* premature end of token */
-      ch = *tok->dptr++;
-      if (ch == '^')
-	mutt_buffer_addch (dest, ch);
-      else if (ch == '[')
-	mutt_buffer_addch (dest, '\033');
-      else if (isalpha ((unsigned char) ch))
-	mutt_buffer_addch (dest, toupper ((unsigned char) ch) - '@');
-      else
-      {
-	mutt_buffer_addch (dest, '^');
-	mutt_buffer_addch (dest, ch);
-      }
-    }
-    else if (ch == '`' && (!qc || qc == '"'))
-    {
-      FILE	*fp;
-      pid_t	pid;
-      char	*cmd, *ptr;
-      size_t	expnlen;
-      BUFFER	expn;
-      int	line = 0;
-
-      pc = tok->dptr;
-      do {
-	if ((pc = strpbrk (pc, "\\`")))
-	{
-	  /* skip any quoted chars */
-	  if (*pc == '\\')
-	    pc += 2;
-	}
-      } while (pc && *pc != '`');
-      if (!pc)
-      {
-        mutt_debug (1, "mutt_get_token: mismatched backticks\n");
-	return (-1);
-      }
-      cmd = mutt_substrdup (tok->dptr, pc);
-      if ((pid = mutt_create_filter (cmd, NULL, &fp, NULL)) < 0)
-      {
-        mutt_debug (1, "mutt_get_token: unable to fork command: %s", cmd);
-	FREE (&cmd);
-	return (-1);
-      }
-      FREE (&cmd);
-
-      tok->dptr = pc + 1;
-
-      /* read line */
-      mutt_buffer_init (&expn);
-      expn.data = mutt_read_line (NULL, &expn.dsize, fp, &line, 0);
-      safe_fclose (&fp);
-      mutt_wait_filter (pid);
-
-      /* if we got output, make a new string consisting of the shell output
-	 plus whatever else was left on the original line */
-      /* BUT: If this is inside a quoted string, directly add output to 
-       * the token */
-      if (expn.data && qc)
-      {
-	mutt_buffer_addstr (dest, expn.data);
-	FREE (&expn.data);
-      }
-      else if (expn.data)
-      {
-	expnlen = mutt_strlen (expn.data);
-	tok->dsize = expnlen + mutt_strlen (tok->dptr) + 1;
-	ptr = safe_malloc (tok->dsize);
-	memcpy (ptr, expn.data, expnlen);
-	strcpy (ptr + expnlen, tok->dptr);	/* __STRCPY_CHECKED__ */
-	if (tok->destroy)
-	  FREE (&tok->data);
-	tok->data = ptr;
-	tok->dptr = ptr;
-	tok->destroy = 1; /* mark that the caller should destroy this data */
-	ptr = NULL;
-	FREE (&expn.data);
-      }
-    }
-    else if (ch == '$' && (!qc || qc == '"') && (*tok->dptr == '{' || isalpha ((unsigned char) *tok->dptr)))
-    {
-      const char *env = NULL;
-      char *var = NULL;
-      int idx;
-
-      if (*tok->dptr == '{')
-      {
-	tok->dptr++;
-	if ((pc = strchr (tok->dptr, '}')))
-	{
-	  var = mutt_substrdup (tok->dptr, pc);
-	  tok->dptr = pc + 1;
-	}
-      }
-      else
-      {
-	for (pc = tok->dptr; isalnum ((unsigned char) *pc) || *pc == '_'; pc++)
-	  ;
-	var = mutt_substrdup (tok->dptr, pc);
-	tok->dptr = pc;
-      }
-      if (var)
-      {
-        if ((env = getenv (var)) || (env = myvar_get (var)))
-          mutt_buffer_addstr (dest, env);
-        else if ((idx = mutt_option_index (var)) != -1)
-        {
-          /* expand settable mutt variables */
-          char val[LONG_STRING];
-
-          if (var_to_string (idx, val, sizeof (val)))
-            mutt_buffer_addstr (dest, val);
-        }
-        FREE (&var);
-      }
-    }
-    else
-      mutt_buffer_addch (dest, ch);
-  }
-  mutt_buffer_addch (dest, 0); /* terminate the string */
-  SKIPWS (tok->dptr);
-  return 0;
 }
 
 static void mutt_free_opt (struct option_t* p)
@@ -3368,7 +3170,7 @@ int mutt_nm_tag_complete (char *buffer, size_t len, int pos, int numtabs)
 }
 #endif
 
-static int var_to_string (int idx, char* val, size_t len)
+int var_to_string (int idx, char* val, size_t len)
 {
   char tmp[LONG_STRING];
   static const char * const vals[] = { "no", "yes", "ask-no", "ask-yes" };
@@ -4161,7 +3963,7 @@ static void myvar_del (const char* var)
   }
 }
 
-static const char* myvar_get (const char* var)
+const char* myvar_get (const char* var)
 {
   myvar_t* cur;
