+++ mutt.h
@@ -97,6 +97,7 @@
 
 #define WHERE_DEFINED 1
 
+#include "buffer.h"
 #include "mutt_regex.h"
 
 /* flags for mutt_enter_string() */
@@ -113,23 +114,6 @@
 #define  MUTT_NM_QUERY (1<<9) /* Notmuch query mode. */
 #define  MUTT_NM_TAG   (1<<10) /* Notmuch tag +/- mode. */
 #endif
-
-/* flags for mutt_get_token() */
-#define MUTT_TOKEN_EQUAL      1       /* treat '=' as a special */
-#define MUTT_TOKEN_CONDENSE   (1<<1)  /* ^(char) to control chars (macros) */
-#define MUTT_TOKEN_SPACE      (1<<2)  /* don't treat whitespace as a term */
-#define MUTT_TOKEN_QUOTE      (1<<3)  /* don't interpret quotes */
-#define MUTT_TOKEN_PATTERN    (1<<4)  /* !)|~ are terms (for patterns) */
-#define MUTT_TOKEN_COMMENT    (1<<5)  /* don't reap comments */
-#define MUTT_TOKEN_SEMICOLON  (1<<6)  /* don't treat ; as special */
-
-typedef struct
-{
-  char *data;	/* pointer to data */
-  char *dptr;	/* current read/write position */
-  size_t dsize;	/* length of data */
-  int destroy;	/* destroy `data' when done? */
-} BUFFER;
 
 /* flags for _mutt_system() */
 #define MUTT_DETACH_PROCESS	1	/* detach subprocess from group */
