+++ pgpkey.c
@@ -25,6 +25,7 @@
 #endif
 
 #include "mutt.h"
+#include "filter.h"
 #include "mutt_curses.h"
 #include "mutt_menu.h"
 #include "mime.h"
