+++ smime.c
@@ -23,6 +23,7 @@
 #endif
 
 #include "mutt.h"
+#include "filter.h"
 #include "mutt_curses.h"
 #include "mutt_menu.h"
 #include "smime.h"
